# update parser. fixes bazingatechnologies/FSharp.Data.GraphQL#2