--- a/tests/FSharp.Data.GraphQL.Tests/ParserTests.fs
+++ b/tests/FSharp.Data.GraphQL.Tests/ParserTests.fs
@@ -602,6 +602,6 @@ fragment frag on Friend {
   query
 }"""
 
-[<Fact(Skip = "Open when new parser will be deployed")>]
+[<Fact>]
 let ``parser should parse kitchen sink``() =
     parse KitchenSink