Update to the parser to handle input objects, fix whitespace issues.