@Horusiath 2016-06-21
I've created benchmarks on my branch (PR #44) after merging a new parser version - white rows are the tests where new parser was included as part of benchmarking:

### Before 

```ini

BenchmarkDotNet=v0.9.7.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU 3.00GHz, ProcessorCount=4
Frequency=2929687 ticks, Resolution=341.3334 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=SimpleExecutionBenchmark  Mode=Throughput  

```
                       Method |      Median |    StdDev |        Mean |         Min |         Max |    Op/s |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
----------------------------- |------------ |---------- |------------ |------------ |------------ |-------- |------- |------ |------ |------------------- |
 BenchmarkSimpleQueryUnparsed | 329.4432 us | 2.3003 us | 329.3496 us | 324.6654 us | 332.9564 us | 3036.29 |  38,61 |     - |     - |           6 221,65 |
   BenchmarkSimpleQueryParsed | 274.8102 us | 3.8274 us | 273.9460 us | 265.6387 us | 281.8587 us | 3650.35 |   8,89 |     - |     - |           2 178,75 |
   BenchmarkFlatQueryUnparsed | 354.8151 us | 2.6875 us | 354.2525 us | 349.1364 us | 360.3077 us | 2822.85 |  64,99 |     - |     - |           9 648,61 |
     BenchmarkFlatQueryParsed | 273.6142 us | 3.2951 us | 274.5027 us | 268.2144 us | 281.3114 us | 3642.95 |   9,09 |     - |     - |           2 223,63 |
 BenchmarkNestedQueryUnparsed | 509.8961 us | 4.5644 us | 511.0484 us | 503.1488 us | 518.8788 us | 1956.76 | 226,00 |     - |     - |          34 556,83 |
   BenchmarkNestedQueryParsed | 277.2645 us | 2.5692 us | 276.5373 us | 270.1634 us | 279.8317 us | 3616.15 |   9,09 |     - |     - |           2 217,65 |

### After

```ini

BenchmarkDotNet-Dev=v0.9.7.0+
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU 3.00GHz, ProcessorCount=4
Frequency=2929687 ticks, Resolution=341.3334 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=SimpleExecutionBenchmark  Mode=Throughput  

```
                       Method |      Median |     StdDev |        Mean |         Min |         Max |    Op/s |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
----------------------------- |------------ |----------- |------------ |------------ |------------ |-------- |------- |------ |------ |------------------- |
 BenchmarkSimpleQueryUnparsed | 262.4884 us |  3.4915 us | 262.8546 us | 258.6030 us | 270.9007 us | 3804.38 |  25.28 |     - |     - |           3 713,42 |
   BenchmarkSimpleQueryParsed | 230.9087 us |  2.1768 us | 230.9260 us | 226.9734 us | 235.7959 us | 4330.39 |  12.28 |     - |     - |           1 975,70 |
   BenchmarkFlatQueryUnparsed | 273.8460 us | 46.4805 us | 293.2557 us | 265.5694 us | 402.1841 us | 3409.99 |  21.00 |     - |     - |           4 167,94 |
     BenchmarkFlatQueryParsed | 234.7738 us |  4.8393 us | 235.4615 us | 229.7132 us | 247.3500 us | 4246.98 |  12.56 |     - |     - |           2 025,43 |
 BenchmarkNestedQueryUnparsed | 327.5806 us |  3.5390 us | 328.1032 us | 321.3334 us | 334.0714 us | 3047.82 | 118.18 |     - |     - |          11 511,98 |
   BenchmarkNestedQueryParsed | 234.1374 us |  2.4245 us | 234.2561 us | 229.1180 us | 240.0487 us | 4268.83 |  13.83 |     - |     - |           2 206,66 |

### Summary

- From 25% to 65% faster execution of the **whole** query (not only parsing phase) - the more complex query, the higher improvement.
- From 30% to almost twice less objects allocated and from almost twice to even trice less total memory allocated.

Remarks:

- Part of benchmark which didn't make use of parser is ~15% faster and uses ~30% more object allocations - mostly because I returned from arrays to lists in some places to keep compatibility with new parser API.